+++ Poker/Node/BetNodes/CallFoldNode.cs
@@ -55,12 +55,12 @@ namespace Poker
             // For each pocket we might have, calculate what we should do.
             for (int p1 = 0; p1 < Pocket.N; p1++)
             {
-                //if (double.IsNaN(PocketP[p1])) continue;
+                //if (decimal.IsNaN(PocketP[p1])) continue;
                 if (!MyCommunity.AvailablePocket[p1]) continue;
 
                 // Get EV for raising/calling/folding.
-                double CallEV = CallBranch.EV[p1];
-                double FoldEV = -Spent;
+                decimal CallEV = CallBranch.EV[p1];
+                decimal FoldEV = -Spent;
 
                 // Decide strategy based on which action is better.
                 if (CallEV >= FoldEV)
@@ -86,16 +86,16 @@ namespace Poker
             // For each pocket we might have, calculate what we expect to happen.
             for (int p1 = 0; p1 < Pocket.N; p1++)
             {
-                //if (double.IsNaN(PocketP[p1])) continue;
+                //if (decimal.IsNaN(PocketP[p1])) continue;
                 if (!MyCommunity.AvailablePocket[p1]) continue;
 
                 // Get likelihoods for opponent raising/calling/folding.
-                double CallChance = TotalChance(PocketP, S, p1);
-                double FoldChance = 1f - CallChance;
+                decimal CallChance = TotalChance(PocketP, S, p1);
+                decimal FoldChance = ((decimal)1) - CallChance;
 
                 // Get EV assuming opponent raising/calling/folding.
-                double CallEV = CallBranch.EV[p1];
-                double FoldEV = Spent;
+                decimal CallEV = CallBranch.EV[p1];
+                decimal FoldEV = Spent;
 
                 // Calculate total EV
                 EV[p1] = CallChance * CallEV +
@@ -104,15 +104,15 @@ namespace Poker
             }
         }
 
-        public override double _Simulate(Var S1, Var S2, int p1, int p2, ref int[] BranchIndex, int IndexOffset)
+        public override decimal _Simulate(Var S1, Var S2, int p1, int p2, ref int[] BranchIndex, int IndexOffset)
         {
             PocketData Data = (ActivePlayer == Player.Button ? S1 : S2)(this);
             int pocket = ActivePlayer == Player.Button ? p1 : p2;
 
-            double CallEV = CallBranch._Simulate(S1, S2, p1, p2, ref BranchIndex, IndexOffset);
-            double FoldEV = ActivePlayer == Player.Button ? -Spent : Spent;
+            decimal CallEV = CallBranch._Simulate(S1, S2, p1, p2, ref BranchIndex, IndexOffset);
+            decimal FoldEV = ActivePlayer == Player.Button ? -Spent : Spent;
 
-            double EV = Data[pocket] * CallEV +
+            decimal EV = Data[pocket] * CallEV +
                         (1 - Data[pocket]) * FoldEV;
             Assert.IsNum(EV);
