+++ Poker/Node/PocketShowdownNode.cs
@@ -32,7 +32,7 @@ namespace Poker
             int PocketValue1, PocketValue2;
             for (int p1 = 0; p1 < Pocket.N; p1++)
             {
-                //if (double.IsNaN(PocketP[p1])) continue;
+                //if (decimal.IsNaN(PocketP[p1])) continue;
                 if (!MyCommunity.AvailablePocket[p1]) continue;
                 PocketValue1 = p1;
 
@@ -41,10 +41,10 @@ namespace Poker
                 UpdateOnExclusion(PocketP, UpdatedP, p1);
                 OpCount++;
 
-                double PocketShowdownEV = 0;
+                decimal PocketShowdownEV = 0;
                 for (int p2 = 0; p2 < Pocket.N; p2++)
                 {
-                    //if (double.IsNaN(UpdatedP[p2])) continue;
+                    //if (decimal.IsNaN(UpdatedP[p2])) continue;
                     if (!MyCommunity.AvailablePocket[p2]) continue;
                     PocketValue2 = p2;
 
@@ -60,7 +60,7 @@ namespace Poker
             }
         }
 
-        public override double _Simulate(Var S1, Var S2, int p1, int p2, ref int[] BranchIndex, int IndexOffset)
+        public override decimal _Simulate(Var S1, Var S2, int p1, int p2, ref int[] BranchIndex, int IndexOffset)
         {
             int PocketValue1 = p1;
             int PocketValue2 = p2;
